fix temp F status in slurm submission

+++ pkg/slurm/Status.go
@@ -199,8 +199,11 @@ func (h *SidecarHandler) StatusHandler(w http.ResponseWriter, r *http.Request) {
 						}
 						resp = append(resp, commonIL.PodStatus{PodName: pod.Name, PodUID: string(pod.UID), PodNamespace: pod.Namespace, Containers: containerStatuses})
 					case "F":
-						if (*h.JIDs)[uid].EndTime.IsZero() {
+						// patch to fix Leonardo temporary F status after submit
+						_, err := os.Stat(path + "/FinishedAt.time")
+						if (*h.JIDs)[uid].EndTime.IsZero() && errors.Is(err, os.ErrNotExist) {
 							(*h.JIDs)[uid].EndTime = timeNow
+
 							f, err := os.Create(path + "/FinishedAt.time")
 							if err != nil {
 								statusCode = http.StatusInternalServerError
@@ -208,17 +211,19 @@ func (h *SidecarHandler) StatusHandler(w http.ResponseWriter, r *http.Request) {
 								return
 							}
 							f.WriteString((*h.JIDs)[uid].EndTime.Format("2006-01-02 15:04:05.999999999 -0700 MST"))
-						}
-						for _, ct := range pod.Spec.Containers {
-							exitCode, err := getExitCode(h.Ctx, path, ct.Name, exitCodeMatch, sessionContextMessage)
-							if err != nil {
-								log.G(h.Ctx).Error(err)
-								continue
+						} else if (*h.JIDs)[uid].EndTime.IsZero() {
+							endTime := timeNow
+							for _, ct := range pod.Spec.Containers {
+								exitCode, err := getExitCode(h.Ctx, path, ct.Name, exitCodeMatch, sessionContextMessage)
+								if err != nil {
+									log.G(h.Ctx).Error(err)
+									continue
+								}
+								containerStatus := v1.ContainerStatus{Name: ct.Name, State: v1.ContainerState{Terminated: &v1.ContainerStateTerminated{StartedAt: metav1.Time{Time: (*h.JIDs)[uid].StartTime}, FinishedAt: metav1.Time{Time: endTime}, ExitCode: exitCode}}, Ready: false}
+								containerStatuses = append(containerStatuses, containerStatus)
 							}
-							containerStatus := v1.ContainerStatus{Name: ct.Name, State: v1.ContainerState{Terminated: &v1.ContainerStateTerminated{StartedAt: metav1.Time{Time: (*h.JIDs)[uid].StartTime}, FinishedAt: metav1.Time{Time: (*h.JIDs)[uid].EndTime}, ExitCode: exitCode}}, Ready: false}
-							containerStatuses = append(containerStatuses, containerStatus)
+							resp = append(resp, commonIL.PodStatus{PodName: pod.Name, PodUID: string(pod.UID), PodNamespace: pod.Namespace, Containers: containerStatuses})
 						}
-						resp = append(resp, commonIL.PodStatus{PodName: pod.Name, PodUID: string(pod.UID), PodNamespace: pod.Namespace, Containers: containerStatuses})
 					case "PD":
 						for _, ct := range pod.Spec.Containers {
 							containerStatus := v1.ContainerStatus{Name: ct.Name, State: v1.ContainerState{Waiting: &v1.ContainerStateWaiting{}}, Ready: false}
